[fix] Adjust UI to adapt to visionOS
Known issue:

The height of the row number is not the set value on visionOS, but is larger than the set value (there is a certain spacing between Cells)

Related questions: [stackoverflow](https://stackoverflow.com/questions/78222887/how-to-remove-the-spacing-around-uitableviewcell-on-visionos)

This issue can cause the collapse and expand buttons to be incorrectly positioned or even completely unusable.

At the same time, the content in this submission is likely caused by this issue. Maybe after we deal with the line number height issue, we will roll back this commit.